news_scraper/backend/services/enrichment-pipeline/requirements.txt
aiohttp
beautifulsoup4
googlenewsdecoder
lxml
numpy
playwright
pymongo
python-Levenshtein
redis
torch
transformers
